added v2 model, improved batching, added capability for parallel mining. improved default scraping parameters

+++ web/models.py
@@ -11,10 +11,7 @@ import torch
 import numpy as np
 import shutil
 from time import time
-from tqdm import trange
-
-
-# from fairseq.models.transformer import TransformerModel
+from tqdm import trange, tqdm
 
 
 def model_setup():
@@ -24,15 +21,20 @@ def model_setup():
                   'https://drive.google.com/u/0/uc?id=195HjThVR1Y0KI93PvLbqpNXw2Iqg4stk&export=download',
               'he_sp_zohar_V1':
                   'https://drive.google.com/u/0/uc?id=1IVPls1YJyJFhrfArKSNkUrZJiQghpT9o&export=download',
+              'he_en_zohar_V2':
+                  'https://drive.google.com/u/0/uc?id=1PHIi0vtHZ00_bzq4tjKhz2ET3AUTwMXK&export=download',
               }
     for model_name, model_url in models.items():
+        is_old = model_name.endswith('V1')
         if not os.path.exists(f'models/{model_name}'):
             output = 'model.zip'
             gdown.download(model_url, output, quiet=False)
             with ZipFile(output, 'r') as zipf:
                 zipf.extractall()
-            shutil.move('content/model', f'models/{model_name}')
-            shutil.rmtree('content')
+            path = 'content'
+            if is_old: path += '/model'
+            shutil.move(path, f'models/{model_name}')
+            if is_old: shutil.rmtree('content')
 
 
 def process(content, lang='he'):
@@ -51,42 +53,46 @@ def split_content(content, max_words, source):
 
 
 class TranslationModel:
-    def __init__(self, model, timestamp, args):
+    def __init__(self, model, timestamp, args, write=True, rel_path='models/'):
+        self.model = model
         parts = model.split('_')
         self.source = parts[0]
         self.target = parts[1]
         self.timestamp = timestamp
-        with open(f'progress/{self.timestamp}.txt', 'w') as f:
-            f.write('0/0')
+        self.write = write
+        if self.write:
+            with open(f'progress/{self.timestamp}.txt', 'w') as f:
+                f.write('0/0')
         self.bs = args.bs
         self.backend = args.backend
         if args.threads != -1:
             torch.set_num_threads(args.threads)
         print(f'Running with {torch.get_num_threads()} threads.')
-        if self.backend == 'huggingface':
-            mname = 'models/' + model
-            self.torch_device = 'cpu'
-            self.trained_model = AutoModelForSeq2SeqLM.from_pretrained(mname).to(self.torch_device)
-            self.trained_tok = AutoTokenizer.from_pretrained(mname)
+        mname = rel_path + model
+        self.torch_device = 'cpu'
+        self.trained_model = AutoModelForSeq2SeqLM.from_pretrained(mname).to(self.torch_device)
+        self.trained_tok = AutoTokenizer.from_pretrained(mname)
 
     def get_translated_text(self, text):
         text = [t for t in text if t]
         trained_translated_txt = ['']
         if not text: return trained_translated_txt
-        if self.backend == 'huggingface':
-            trained_batch = self.trained_tok(text, return_tensors='pt', padding=True).to(self.torch_device)
-            assert len(trained_batch['input_ids'][0]) < 512, 'Tokenized batch too long!'
-            trained_translated = self.trained_model.generate(**trained_batch, num_beams=1, early_stopping=False)
-            trained_translated_txt = self.trained_tok.batch_decode(trained_translated, skip_special_tokens=True)
-        # elif self.backend == 'fairseq':
-        #     trained_translated_txt = self.trained_model.translate(text)
+        trained_batch = self.trained_tok(text, return_tensors='pt', padding=True).to(self.torch_device)
+        assert len(trained_batch['input_ids'][0]) < 512, 'Tokenized batch too long!'
+        trained_translated = self.trained_model.generate(**trained_batch, num_beams=1)
+        trained_translated_txt = self.trained_tok.batch_decode(trained_translated, skip_special_tokens=True)
         return trained_translated_txt
 
     def translate(self, s):
         bs = self.bs
-        t = time()
+        tm = time()
         res_arr = []
-        batch_s = s.split('\n')
+        batch_s = np.array([t for t in s.split('\n') if t])
+        lns = np.array([len(t.split()) for t in batch_s])
+        # argsort for batching and then invert it
+        argsort1 = lns.argsort()
+        argsort2 = argsort1.argsort()
+        batch_s = batch_s[argsort1]
         if bs != -1:
             n_batches = int(np.ceil(len(batch_s) / bs))
             for b in trange(n_batches):
@@ -94,19 +100,22 @@ class TranslationModel:
                 end = start + bs
                 curr_batch = batch_s[start:end]
                 res_arr.extend(self.get_translated_text(curr_batch))
-                with open(f'progress/{self.timestamp}.txt', 'w') as f:
-                    f.write(f'{b + 1}/{n_batches}')
+                if self.write:
+                    with open(f'progress/{self.timestamp}.txt', 'w') as f:
+                        f.write(f'{b + 1}/{n_batches}')
         else:
             res_arr = self.get_translated_text(s)
+        res_arr = np.array(res_arr)[argsort2]
         result = '\n'.join(res for res in res_arr)
-        print(f'Time taken: {time() - t}, batch size: {bs}')
+        print(f'Time taken: {time() - tm}, batch size: {bs}')
         return result
 
     def __call__(self, mimetype, content):
         res = {'target': 'Translation 1', 'source': 'Not a text file'}
         if mimetype == 'application/vnd.openxmlformats-officedocument.wordprocessingml.document':
             content = process(content)
-        split_txt = split_content(content, max_words=200, source=self.source)
+        max_words = 400 if self.model.endswith('V2') else 200
+        split_txt = split_content(content, max_words=max_words, source=self.source)
         txt = join_text(split_txt)
         translated_txt = self.translate(txt)
         txt = '\n\n'.join(txt.split('\n'))
@@ -117,4 +126,5 @@ class TranslationModel:
         return res
 
 
-model_setup()
+if __name__ == '__main__':
+    model_setup()
